Fix printf format warnings for 64-bit build.
Fix warnings caused by size_t and ACPI_SIZE changing to 64 bits.

+++ source/components/namespace/nsnames.c
@@ -241,7 +241,7 @@ AcpiNsGetExternalPathname (
     NameBuffer = ACPI_ALLOCATE_ZEROED (Size);
     if (!NameBuffer)
     {
-        ACPI_ERROR ((AE_INFO, "Could not allocate %d bytes", Size));
+        ACPI_ERROR ((AE_INFO, "Could not allocate %u bytes", (UINT32) Size));
         return_PTR (NULL);
     }
